Rename pytorch_dataloader.py to dataset.py because it makes more sense
